+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/wfReactInterface/wfDataModelHooksInterface.ts
@@ -12,8 +12,8 @@
 import * as Types from '../../../../../../core/model/types';
 import {WeaveKind} from '../../../../../../react';
 import {KNOWN_BASE_OBJECT_CLASSES, OP_CATEGORIES} from './constants';
-import * as traceServerClient from './traceServerClient'; // TODO: This import is not ideal, should delete this whole interface
 import {Query} from './traceServerClientInterface/query'; // TODO: This import is not ideal, should delete this whole interface
+import * as traceServerClientTypes from './traceServerClientTypes'; // TODO: This import is not ideal, should delete this whole interface
 
 export type OpCategory = (typeof OP_CATEGORIES)[number];
 export type KnownBaseObjectClassType =
@@ -47,7 +47,7 @@ export type CallSchema = CallKey & {
   rawFeedback?: any;
   userId: string | null;
   runId: string | null;
-  traceCall?: traceServerClient.TraceCallSchema; // this will eventually be the entire call schema
+  traceCall?: traceServerClientTypes.TraceCallSchema; // this will eventually be the entire call schema
 };
 
 export type CallFilter = {
@@ -164,7 +164,7 @@ export type WFDataModelHooksInterface = {
     filter: CallFilter,
     limit?: number,
     offset?: number,
-    sortBy?: traceServerClient.SortBy[],
+    sortBy?: traceServerClientTypes.SortBy[],
     query?: Query,
     expandedRefColumns?: Set<string>,
     opts?: {skip?: boolean; refetchOnDelete?: boolean}
@@ -175,7 +175,7 @@ export type WFDataModelHooksInterface = {
     filter: CallFilter,
     query?: Query,
     opts?: {skip?: boolean; refetchOnDelete?: boolean}
-  ) => Loadable<traceServerClient.TraceCallsQueryStatsRes>;
+  ) => Loadable<traceServerClientTypes.TraceCallsQueryStatsRes>;
   useCallsDeleteFunc: () => (
     entity: string,
     project: string,
