+++ weave-js/src/components/PagePanelComponents/Home/Browse3/types.ts
@@ -1,6 +1,6 @@
 import {GridColDef, GridColumnGroupingModel} from '@mui/x-data-grid-pro';
 
-import {TraceCallSchema} from './pages/wfReactInterface/traceServerClient';
+import {TraceCallSchema} from './pages/wfReactInterface/traceServerClientTypes';
 
 export type ColumnInfo = {
   cols: Array<GridColDef<TraceCallSchema>>;
